initial refactor to DD menu
