+++ gamestonk_terminal/stocks/due_diligence/dd_controller.py
@@ -7,19 +7,20 @@ from typing import List
 from pandas.core.frame import DataFrame
 from prompt_toolkit.completion import NestedCompleter
 
-from gamestonk_terminal.stocks.due_diligence import business_insider_view as bi_view
 from gamestonk_terminal.stocks.due_diligence import (
-    financial_modeling_prep_view as fmp_view,
+    financial_modeling_prep_view,
+    business_insider_view,
+    finviz_view,
+    market_watch_view,
+    quandl_view,
+    news_view,
+    finra_view,
+    sec_view,
+    stockgrid_dd_view,
+    finnhub_view,
+    csimarket_view,
 )
-from gamestonk_terminal.stocks.due_diligence import finviz_view as fvz_view
-from gamestonk_terminal.stocks.due_diligence import market_watch_view as mw_view
-from gamestonk_terminal.stocks.due_diligence import quandl_view as q_view
-from gamestonk_terminal.stocks.due_diligence import news_view
-from gamestonk_terminal.stocks.due_diligence import finra_view
-from gamestonk_terminal.stocks.due_diligence import sec_view
-from gamestonk_terminal.stocks.due_diligence import stockgrid_dd_view as sg_view
-from gamestonk_terminal.stocks.due_diligence import finnhub_view
-from gamestonk_terminal.stocks.due_diligence import csimarket_view
+
 from gamestonk_terminal import feature_flags as gtff
 from gamestonk_terminal.helper_funcs import get_flair
 from gamestonk_terminal.menu import session
@@ -40,8 +41,6 @@ class DueDiligenceController:
         "pt",
         "rot",
         "est",
-        "ins",
-        "insider",
         "news",
         "analyst",
         "sec",
@@ -110,8 +109,6 @@ class DueDiligenceController:
         print(
             "   est           quarter and year analysts earnings estimates [Business Insider]"
         )
-        print("   ins           insider activity over time [Business Insider]")
-        print("   insider       insider trading of the company [Finviz]")
         print("   sec           SEC filings [Market Watch]")
         print("   short         short interest [Quandl]")
         print("   dp            dark pools (ATS) vs OTC data [FINRA]")
@@ -166,21 +163,17 @@ class DueDiligenceController:
         """Process Quit command - quit the program"""
         return True
 
-    def call_insider(self, other_args: List[str]):
-        """Process insider command"""
-        fvz_view.insider(other_args, self.ticker)
-
     def call_news(self, other_args: List[str]):
         """Process news command"""
         news_view.news(other_args, self.ticker)
 
     def call_analyst(self, other_args: List[str]):
         """Process analyst command"""
-        fvz_view.analyst(other_args, self.ticker)
+        finviz_view.analyst(other_args, self.ticker)
 
     def call_pt(self, other_args: List[str]):
         """Process pt command"""
-        bi_view.price_target_from_analysts(
+        business_insider_view.price_target_from_analysts(
             other_args, self.stock, self.ticker, self.start, self.interval
         )
 
@@ -190,25 +183,19 @@ class DueDiligenceController:
 
     def call_est(self, other_args: List[str]):
         """Process est command"""
-        bi_view.estimates(other_args, self.ticker)
-
-    def call_ins(self, other_args: List[str]):
-        """Process ins command"""
-        bi_view.insider_activity(
-            other_args, self.stock, self.ticker, self.start, self.interval
-        )
+        business_insider_view.estimates(other_args, self.ticker)
 
     def call_rating(self, other_args: List[str]):
         """Process rating command"""
-        fmp_view.rating(other_args, self.ticker)
+        financial_modeling_prep_view.rating(other_args, self.ticker)
 
     def call_sec(self, other_args: List[str]):
         """Process sec command"""
-        mw_view.sec_fillings(other_args, self.ticker)
+        market_watch_view.sec_fillings(other_args, self.ticker)
 
     def call_short(self, other_args: List[str]):
         """Process short command"""
-        q_view.short_interest(other_args, self.ticker, self.start)
+        quandl_view.short_interest(other_args, self.ticker, self.start)
 
     def call_dp(self, other_args: List[str]):
         """Process dp command"""
@@ -220,11 +207,11 @@ class DueDiligenceController:
 
     def call_shortview(self, other_args: List[str]):
         """Process shortview command"""
-        sg_view.shortview(self.ticker, other_args)
+        stockgrid_dd_view.shortview(self.ticker, other_args)
 
     def call_darkpos(self, other_args: List[str]):
         """Process darkpos command"""
-        sg_view.darkpos(self.ticker, other_args)
+        stockgrid_dd_view.darkpos(self.ticker, other_args)
 
     def call_supplier(self, other_args: List[str]):
         """Process supplier command"""
