+++ gamestonk_terminal/stocks/stocks_controller.py
@@ -228,7 +228,12 @@ What do you want to do?
 
     def call_ins(self, _):
         """Process ins command"""
-        ret = insider_controller.menu()
+        ret = insider_controller.menu(
+            self.ticker,
+            self.start,
+            self.interval,
+            self.stock,
+        )
         if ret is False:
             self.print_help()
         else:
